pyproject.toml
[project]
name = "europeana-qlever"
version = "0.1.0"
description = "CLI for ingesting Europeana EDM TTL into QLever and exporting to Parquet"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "rich>=14.3.3",
    "rdflib>=7.6.0",
    "httpx>=0.28.1",
    "duckdb>=1.5.1",
    "pyarrow>=23.0.1",
    "psutil>=7.0.0",
    "linkml-runtime>=1.8.0",
    "mlcroissant>=1.0",
    "pyyaml>=6.0",
    "websockets>=16.0",
]

[project.scripts]
europeana-qlever = "europeana_qlever.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/europeana_qlever"]

[dependency-groups]
dev = []

